separate camera controller
