+++ examples/tools/scene_viewer/main.rs
@@ -14,13 +14,15 @@ use bevy::{
     window::WindowPlugin,
 };
 
+#[path = "../../helpers/camera_controller.rs"]
+mod camera_controller;
+
 #[cfg(feature = "animation")]
 mod animation_plugin;
-mod camera_controller_plugin;
 mod morph_viewer_plugin;
 mod scene_viewer_plugin;
 
-use camera_controller_plugin::{CameraController, CameraControllerPlugin};
+use camera_controller::{CameraController, CameraControllerPlugin};
 use morph_viewer_plugin::MorphViewerPlugin;
 use scene_viewer_plugin::{SceneHandle, SceneViewerPlugin};
